fix(ci): python-full-coverage line limit — oversized executor error-path shard exceeded maintainability gate
