+++ plans/PR292_LINE_LIMIT_VALIDATION.md
@@ -1,0 +1,38 @@
+# PR292 Line Limit Validation
+
+Plan reference: `plans/PR292_LINE_LIMIT_PLAN.md`
+
+## Requirement status
+
+- Keep the line-limit check unchanged: Complete.
+- Split the oversized test file so every first-party file is below 1,500 lines: Complete.
+- Preserve pytest discovery and behavior for the moved tests: Complete.
+- Run focused validation only; full AWF/GitHub validation remains managed by AWF after agent completion: Complete.
+- Commit the fix locally without switching branches or pushing: Complete by the local commit containing this validation document.
+
+## Evidence
+
+Files changed:
+
+- `tests/unit/control/test_executor_error_paths_parts/test_executor_error_paths_part_006.py`
+- `tests/unit/control/test_executor_error_paths_parts/test_executor_error_paths_part_010.py`
+- `plans/PR292_LINE_LIMIT_PLAN.md`
+- `plans/PR292_LINE_LIMIT_VALIDATION.md`
+
+Focused checks run:
+
+- `uv run --python 3.12 --extra dev pytest tests/unit/test_core_decomposition_maintainability.py::test_first_party_code_files_stay_under_line_limit -q` — passed.
+- `uv run --python 3.12 --extra dev pytest tests/unit/control/test_executor_error_paths_parts/test_executor_error_paths_part_010.py -q` — passed.
+- `uv run --python 3.12 --extra dev pytest tests/unit/control/test_executor_error_paths_parts/test_executor_error_paths_part_006.py -q` — passed.
+- `uv run --python 3.12 --extra dev ruff check tests/unit/control/test_executor_error_paths_parts/test_executor_error_paths_part_006.py tests/unit/control/test_executor_error_paths_parts/test_executor_error_paths_part_010.py` — passed.
+
+Line counts after split:
+
+- `test_executor_error_paths_part_006.py`: 1,380 lines.
+- `test_executor_error_paths_part_010.py`: 310 lines.
+
+Full AWF/GitHub validation was not run locally because the workspace contract assigns broad validation, coverage gates, and merge gating to AWF after agent completion.
+
+## Gaps
+
+None.
